moved update tags to hook, added try catch
